Update README.md
Zuber is a new ride-sharing company launching in Chicago. As a data analyst, my goal is to identify patterns in the available data to better understand passenger preferences and the impact of external factors, such as weather, on ride frequency.

This project involves working with a database containing information on taxi trips in Chicago, analyzing competitor data, and testing a hypothesis regarding the effect of weather conditions on the number of rides.